requirements.txt
selenium
python-dateutil
geopy

